Move zenml.new.steps.log_artifact_metadata > zenml.artifacts.utils
